feat: add comment property, use advanced property (#40)
## Which problem is this PR solving?

- TemplateComponents now have a Comment property
- Several of the common components now have some of their properties
marked with the Advanced flag

The advanced flag is an indication that most configurations will not
need to modify that property, so it is hidden by default in the UI under
a control that looks like "> Advanced". Clicking that control will
reveal all those properties. It's intended to allow components to have a
lot of configurability but to hide it from the average user.

## Short description of the changes

- Add comment field to TemplateComponent
- delete unused function that was mistakenly committed
- Mark some properties as Advanced