fix: Translator generates valid refinery rules when no sampler component is provided (#54)
## Which problem is this PR solving?

When generating configs, if no sampler component is provided the
translator does not produce a valid refinery rules configuration (no
default sampler).

This PR updates the sampler to check if a sampler was provided in the
list of components, and if not, adds a default deterministic sampler
with a sample rate of 1.

## Short description of the changes

- Add style property to hpsf.Component
- Add `maybeAddDefaultSampler` that checks if there the components
contains a sampler and adds one if not
- Add unit test