feat: change the way rules files are generated (#90)
## Which problem is this PR solving?

- Rules files needed to be changed so that they're composable. The old
way didn't work. This changes how they're generated and adds some new
semantics so that we can build useful rules components. The problem is
in the weirdly nested nature of refinery rules -- I wanted to avoid a
structure that was tightly tied to refinery's current configuration
structure but is still flexible enough to handle it. It's definitely got
some rather refinery-specific code in here but it's not quite as brittle
as an explicit structure.

## Short description of the changes

- Add a rules_config config type (it uses dotted config but has some
rules-specific fields and has extra semantics)
- Add a compact way of expressing rules conditions and a special way to
handle them in the rules generator
- Add the ability to specify an index so we can create multiple rows
- Add a single rules-specific component
- Tweak the existing samplers to use the new rules_config
- Write some tests 
- Fix up some existing tests (for a couple of tests driven by control
files, I added code to allow rewriting the control files within the
test).
- Add some docs