fix: TraceConverter default values (#92)
## Which problem is this PR solving?

- fixes https://github.com/honeycombio/pipeline-team/issues/383

We found the following issues with the traceconverter default install:
1. it uses the `{STRAWS_REFINERY_POD_IP}` env var that the collector
does not have access to.
2. Sets default otlp http port to 4318, but refinery listens for otlp
http on port 80
3. Configures the otlphttp exporter to communicate using TLS, but
Refinery is not using TLS.
4. The property is named Host, which implies a schema (http/https) does
not need to be provided, but the default does not provide it.

## Short description of the changes

- Updated env var to `${STRAWS_REFINERY_SERVICE}`. We'll update the helm
chart to set this env var using the refinery service name it generates:
https://github.com/honeycombio/helm-charts/pull/488
- update default port to 80
- update default TLS to be insecure. We changed the property name to be
`Secure` instead of `Insecure` so a default value of `false` meant that
the exporter would be rendered with `insecure: true`. We needed to
inverted it since setting a boolean property to `false` via the HPSF
config is not possible; the engine sees that as "not set".
- update rendered endpoint to set `http`/`https` based on `Secure`
property.