feat: Simplify using default values in components (#94)
## Which problem is this PR solving?

- In order to handle defaults, we were relying on a magic incantation in
components: `firstNonzero v1 v2 v3` with a particular pattern of values
based on default values, user-specified values, and values coming from
the environment. This PR replaces all of that with a single helper
method on the TemplateComponent, `Values()`, which does all of that work
in one reference.
- It also removes the now-obsolete `firstNonzero`, because if we need it
again, it turns out it's pretty much just the Go template's `or`
operator.

## Short description of the changes

- Add Values method
- Replace firstNonzero references in a bunch of components
- Remove firstNonzero function
-