test: Added end to end framework for config rendering (#106)
## Which problem is this PR solving?

The HPSF rendering only checks for valid YAML, but not valid configs.
The smoke tests can and will get long to run, this is a much faster to
way to achieve the same goal.

## Short description of the changes

Added a separate go module that encompasses the tests to ensure that
dependencies don't get included.