fix: update emathroughput template to use updated fields (#114)
## Which problem is this PR solving?

- closes https://github.com/honeycombio/pipeline-team/issues/442

## Short description of the changes

- Update to use new fields like Input -> Traces, Output -> Events,
Honeycomb -> HoneycombEvents, Kept -> SampledOutput

To test the config is valid, I added an API key to the HoneycombExporter
and ran `make smoke`. I might try to update that in a separate PR to add
that when testing, because if we were running this in CI we would catch
this.

```yaml
  - name: Honeycomb Exporter 1
    kind: HoneycombExporter
    properties:
      - name: APIKey
        value: hccik_01jj2jj42424jjjjjjj2jjjjjj424jjj2jjjjjjjjjjjjjjj4jjjjj24jj # its fake no worries
```