fix: run smoke tests for templates in ci
* honeycomb exporter requires an API key
in order to run. that is the only thing
preventing these templates from running
successfully in a smoke test, so set a
fake api key in the test to ensure everything else works as expected.
* This still only checks templates, not all
components, so it is still quick to run but
adds a ton of confidence for templates.