fix: make it easier to run smoke tests for templates (#115)
## Which problem is this PR solving?

- template was broken and could've been caught sooner

## Short description of the changes

* honeycomb exporter requires an API key in order to run. that is the
only thing preventing these templates from running
successfully in a smoke test, so set a fake api key in the test to
ensure everything else works as expected.
* This still only checks templates, not all components, so it is still
quick to run but adds a ton of confidence for templates.

This would be nice to run in CI but in a quick timebox found it needed a
few more things to be successful (install yq for instance). May revisit
later but in the meantime this is still useful to see "will it run"