fix: minor updates to component names & properties (#151)
## Which problem is this PR solving?

While building a pipeline, I accidentally pulled in the OTel HTTP
exporter because it was called "Send to Honeycomb". This should be the
name of the Honeycomb exporter. So I went into to make that update, and
made a few more related ones while I was here.

## Short description of the changes

- Renamed OTelHTTPExporter from "Send to Honeycomb" to "Send OTel via
HTTP"
- Renamed OTelGRPCExporter from "OTel gRPC Exporter" to "Send OTel via
gRPC" (to match)
- Renamed HoneycombExporter from "Honeycomb Exporter" to "Send to
Honeycomb"
- Hid the Nop receiver and exporter, because they don't guide users
enough to be valuable right now.
- Moved the Headers property in the StartSampling component into
Advanced, as discussed in a previous Slack thread.
- Made some minor updates to summaries and descriptions.

---------

Co-authored-by: Alex Boten <223565+codeboten@users.noreply.github.com>