fix: automatically set otlp http exporter scheme based on insecure (#153)
## Which problem is this PR solving?

- We updated the validators to enforce that the Host property is a host
or IP, like we wanted, but we also needed to update the template to set
the scheme since it was no longer settable via the Host property.

The otlp http exporter now sets the http scheme based on the `insecure`
property. This follows the same strategy the StartSampling component
uses.

## Short description of the changes

- Update OTLP HTTP exporter component to set an http scheme based on
Insecure propty.
- Update template to use `.Values`.
- Update tests.