feat: rewrite the way refinery samplers work (#159)
# THIS IS A BREAKING CHANGE TO EXISTING HPSF FILES

I'm really sorry this couldn't be smaller, but it couldn't.

## Which problem is this PR solving?

Sampler configurations created with previous versions of HTP weren't
composable -- you couldn't do interesting sampler stuff. So we did some
research, and came up with a new paradigm. This PR implements it.

Big picture changes:
* Samplers are connected with a new type of connection -- SampleData.
It's currently gray because we haven't done anything better in the UI
yet.
* StartSampling now begins a sampling chain. It has 4 outputs, and they
are evaluated in order.
* Each chain consists of 0 or more conditions, each of which is ANDed
together. If a trace being evaluated matches a condition, it moves on to
the next.
* Each chain ends with a sampler. The sampler only decides on the
probability of keeping that trace. We currently have keep, drop,
deterministic and a couple of ema samplers. We can create more.
* If a condition fails, the trace moves on to the next output in
StartSampling. We want to evaluate how that component should work (how
many outputs should it have) but that's outside this discussion.
* Samplers (except Drop) then have an output that goes to the exporter.
* Lots of tests got changed/fixed. A couple of obsolete tests were
deleted.
* Validations are not yet complete. There are things that should be
checked that are not yet. We'll do that separately.

This system implements the ability for StartSampling to specify the
environment name, to support multiple sets of rules the way Refinery
does, but that's blocked by other work in HTP for now, so we don't use
it yet.

## Still to do in future PRs:

- Consider whether some of these components or ports should have
different names
- Enhance validations to include the new sampler rules
- build the rest of the sampler components
- build a bunch more condition components, including a generic one
- fix the readme documentation
- Add more tests for more sampler combinations
- Add a couple of scenario tests
- Generate the Name field in Refinery rules in a better way; today it's
just the name of the sampler
- Dynamic handle creation in StartSampling, or multiple StartSampler
components with different numbers of handles.
- Add env property to StartSampling (also requires other work to use it
properly)