feat: add validation for component version
This updates the validation path to check that any component version
specified in the HPSF will be taken into consideration when looking
up the component. The component map should probably be a map of maps
to include multiple versions of the same component in the near future, but
this is ok for now.

Signed-off-by: alex boten <223565+codeboten@users.noreply.github.com>