chore: clean up following linter suggestions
Ran a linter locally against the repo and am following some of the
suggestions to clean up the code a bit. No functional changes in
this PR, though it did find a potential bug in main.go where a file
handle could be left dangling, and simplified some of the indented
cases that were not necessary.

Signed-off-by: alex boten <223565+codeboten@users.noreply.github.com>