feat: Add autolayout of HPSF (#239)
## Which problem is this PR solving?

This adds an autolayout feature to the HPSF module. I tried several
different algorithms for autolayout in the browser, and none of them
worked well because there is no library I could find in either Go or
JS/TS that does smart graph layout with multiple ports per node.

So I had to write one. Fortunately, our design is simpler than most
layout engines need to contend with, so it was possible to write down a
detailed list of specifications for the design (see the top of
layout.go). I coached an AI step-by-step through creating this, and
eventually got there successfully. I added a "layout" command to the
`hpsf` tool for testing.

Note that this requires adding size information to the `layout` section
of the HPSF in the UI, because our nodes are variable size and the size
is determined by the UI. Once it seemed to be working, I tested it by
adding just that feature in a local build and using copy, the command
line tool, and paste to see the results onscreen.

This algorithm is heuristic and won't always work perfectly (it *can't*
always work because people can build weird stuff). But it handles all
the cases I threw at it. It's not as good at perfect alignment as the
previous algorithm, but it is much better at not having crossed wires.
It should make things better even if perfect is not attainable.

This PR is paired [with one that calls this API from within the hound
UI](https://github.com/honeycombio/hound/pull/27701).

Yes, this is big, but:
* Lots of the changes are detailed tests
* Partial PRs aren't particularly helpful since this system wasn't
useful or functional until substantially all of it was functional.

## Short description of the changes

- Add `layout` package to HPSF
- Implement and document layout algorithm
- Make the cmd line tool use it
- Add lots and lots of tests for various bits of the algorithm