feat: Add Inspect func to Translator to get component info [PIPE-137] (#245)
## Which problem is this PR solving?

Adds a new translator function `Inspect` to get all components in a HPSF
definition. The returned type, `InspectionResult` includes a list of all
components and provides filtering options. Components are are a
`ComponentInfo` struct that includes it's name, style, kind and property
map, taking into account component defaults and user provided values.

The supported filters are:
- Exporters
- Processors
- Receivers
- Samplers (contains dropper, condition, sampler and startsampling)

## Short description of the changes

- Add new InspectionResult Predicate structs and funcs
- Add unit tests to verify behaviour including getting all components
and using filtering

---------

Co-authored-by: Tyler Helmuth <12352919+TylerHelmuth@users.noreply.github.com>
Co-authored-by: Kent Quirk <kentquirk@honeycomb.io>