feat: support VersionError when rendering config (#257)
## Which problem is this PR solving?

Today if a render fails because of an artifact version or component
version we return an error, but callers have no way to identify if the
returned error is because of something internal (a real error) or
because the versions are incompatible. This PR adds a new VersionError
type that callers of GenerateConfig can use to check if the returned
error was because of a version mismatch.

## Short description of the changes

- add new VersionError type
- update component and artifact checks to use the new error type.